Moved PlayerCommands to Player.Commands
